feat: Add userAtom to store user login information
This commit adds the userAtom to store the user's login information. It imports the User type from "@/types/user" and the loadable function from "jotai/utils". The userAtom is used to store the user's login information, and it is initialized as null. This atom will be used to manage the user's login state throughout the application.

Co-authored-by: [Author Name]